Robot and Server coding finished
Full communication between server and robot. Server can send movement
commands to the robot. This can be done from any client connected to the
server, however functionality may be added to allow for this to be done
from the server as well. - Currently no need as can be done from
index.js repl